Remove ObjList*/ObjDict* free functions, rename Float to Double
This commit simplifies the API by removing the ObjList* and ObjDict*
free functions in favor of users working directly with Go-native types.
It also fixes the Float/Double naming inconsistency.

Architecture decisions:
- Users now call obj.List() to get []*Obj and work with Go slice operations
- Users now call obj.Dict() to get *DictType with Items map and Order slice
- Renamed i.Float() to i.Double() for consistency with obj.Double() and DoubleType
- DictKV and DictFrom now build Go maps directly instead of using ObjDictSet

What was removed:
- ObjListLen, ObjListAt, ObjListAppend, ObjListSet (from interp_convert.go)
- ObjDictGet, ObjDictSet, ObjDictKeys, ObjDictLen (from interp_convert.go)

What was updated:
- feather.go: Float→Double, DictKV/DictFrom build maps directly
- interp_callbacks.go: goListSetAt, goDictGet, goDictSet use direct manipulation
- interp_foreign.go: listAppendObj, dictSetObj, dictGetObj use direct manipulation
- cmd/feather-httpd: Build Go slices/maps first, then create objects
- api_test.go, feather_test.go: Updated to use new API

Resulting API is now consistent:
| Constructor    | Conversion      | Type       |
|----------------|-----------------|------------|
| i.Int(42)      | obj.Int()       | IntType    |
| i.Double(3.14) | obj.Double()    | DoubleType |
| i.String("hi") | obj.String()    | (string)   |
| i.List(...)    | obj.List()→[]*Obj | ListType |
| i.Dict()       | obj.Dict()→*DictType | DictType |

Current state: All 1529 harness tests pass. API is cleaner with fewer
ways to do the same thing.

Next steps for colleague:
1. Consider documenting the new API patterns in CLAUDE.md or README
2. The internal handle methods (newIntObj, newDoubleObj, etc.) remain
   for C interop but are unexported
3. All external API now uses Go-native types for maximum interoperability

+++ feather_test.go
@@ -323,8 +323,8 @@ func TestInterpObjectCreation(t *testing.T) {
 		}
 	})
 
-	t.Run("Float", func(t *testing.T) {
-		v := interp.Float(3.14)
+	t.Run("Double", func(t *testing.T) {
+		v := interp.Double(3.14)
 		if v.Type() != "double" {
 			t.Errorf("expected type 'double', got %q", v.Type())
 		}
